requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0

SQLAlchemy==2.0.32
psycopg2-binary

python-dotenv==1.0.1
pydantic
pydantic-settings

PyJWT==2.8.0
